refactor(script): split item buildins into buildin_item.c
Issue #61 second step: move inventory, equip, refine, and card helpers out of script.c and share refine_posword via script_internal.h.
