+++ vcproj/map-server.vcxproj
@@ -207,6 +207,7 @@
     <ClCompile Include="..\src\map\quest.c" />
     <ClCompile Include="..\src\map\ranking.c" />
     <ClCompile Include="..\src\map\script.c" />
+    <ClCompile Include="..\src\map\buildin_item.c" />
     <ClCompile Include="..\src\map\buildin_map.c" />
     <ClCompile Include="..\src\map\skill.c" />
     <ClCompile Include="..\src\map\skill_castend_damage.c" />
