+++ vcproj/map-server.vcxproj.filters
@@ -103,6 +103,9 @@
     <ClCompile Include="..\src\map\script.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_item.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\buildin_map.c">
       <Filter>map_sql</Filter>
     </ClCompile>
